+++ common/src/main/java/com/getwemap/example/common/Constants.kt
@@ -7,6 +7,4 @@ object Constants {
     const val token: String = "GUHTU6TYAWWQHUSR5Z5JZNMXX" // TODO: Modify this value with your own wemap token
 
     val polestarApiKey: String get() = throw Exception("Specify polestarApiKey and remove exception")
-
-    val vpsEndpoint: String get() = throw Exception("Specify vpsEndpoint and remove exception")
 }