+++ positioning/src/main/java/com/getwemap/example/positioning/Config.kt
@@ -1,0 +1,24 @@
+package com.getwemap.example.positioning
+
+import android.content.Context
+import androidx.preference.PreferenceManager
+import com.getwemap.sdk.core.CoreConstants
+
+object Config {
+
+    fun applyGlobalOptions(context: Context) {
+        val prefs = PreferenceManager.getDefaultSharedPreferences(context)
+        with(CoreConstants) {
+            USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED = prefs
+                .getBoolean(PreferenceKey.USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED.name, true)
+            USER_LOCATION_PROJECTION_ON_GRAPH_ENABLED = prefs
+                .getBoolean(PreferenceKey.USER_LOCATION_PROJECTION_ON_GRAPH_ENABLED.name, false)
+        }
+    }
+}
+
+enum class PreferenceKey {
+    // Global
+    USER_LOCATION_PROJECTION_ON_ITINERARY_ENABLED,
+    USER_LOCATION_PROJECTION_ON_GRAPH_ENABLED
+}